Create BaseQuantizer
